Better README and better performance
Updated the README to provide a summary of the two modules up front with
Big O analysis of each. Replaced v2's use of a TreeMultiset with a
HashMultiset to improve the performance characteristics (given the
likely data sizes involved, this makes sense). Added Java debug to the
command line launcher.